requirements.txt
Flask
Flask-Script
flake8
gunicorn
requests
Flask-SQLAlchemy
Flask-Migrate
Flask-API
alembic
psycopg2
pycrypto
